minor: moved location of BaseLinearOperator
