+++ pylops/avo/avo.py
@@ -16,7 +16,7 @@ import numpy as np
 import numpy.typing as npt
 from numpy import cos, sin, tan
 
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils._internal import _value_or_sized_to_tuple
 from pylops.utils.backend import get_array_module
 from pylops.utils.decorators import reshaped
