+++ pylops/basicoperators/memoizeoperator.py
@@ -5,7 +5,7 @@ from typing import List, Tuple
 import numpy as np
 
 from pylops import LinearOperator
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils.typing import NDArray
