+++ pylops/basicoperators/pad.py
@@ -4,7 +4,7 @@ from typing import Sequence, Tuple, Union
 
 import numpy as np
 
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils._internal import _value_or_sized_to_tuple
 from pylops.utils.decorators import reshaped
 from pylops.utils.typing import DTypeLike, InputDimsLike, NDArray
