+++ pylops/basicoperators/regression.py
@@ -5,7 +5,7 @@ import logging
 import numpy as np
 import numpy.typing as npt
 
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils.backend import get_array_module
 from pylops.utils.typing import DTypeLike, NDArray
