+++ pylops/signalprocessing/chirpradon2d.py
@@ -4,9 +4,10 @@ import logging
 
 import numpy as np
 
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils.decorators import reshaped
 from pylops.utils.typing import DTypeLike, NDArray
+
 from ._chirpradon2d import _chirp_radon_2d
 
 logging.basicConfig(format="%(levelname)s: %(message)s", level=logging.WARNING)
