+++ pylops/signalprocessing/chirpradon3d.py
@@ -4,10 +4,11 @@ import logging
 
 import numpy as np
 
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils import deps
 from pylops.utils.decorators import reshaped
 from pylops.utils.typing import DTypeLike, NDArray
+
 from ._chirpradon3d import _chirp_radon_3d
 
 pyfftw_message = deps.pyfftw_import("the chirpradon3d module")
