+++ pylops/signalprocessing/dwt.py
@@ -7,7 +7,7 @@ from typing import Union
 import numpy as np
 
 from pylops.basicoperators import Pad
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.utils import deps
 from pylops.utils._internal import _value_or_sized_to_tuple
 from pylops.utils.typing import DTypeLike, InputDimsLike, NDArray
