+++ pylops/waveeqprocessing/blending.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from pylops import LinearOperator
 from pylops.basicoperators import BlockDiag, HStack, Pad
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.signalprocessing import Shift
 from pylops.utils.backend import get_array_module
 from pylops.utils.decorators import reshaped
