+++ pylops/waveeqprocessing/kirchhoff.py
@@ -8,7 +8,7 @@ from typing import Optional, Tuple, Union
 
 import numpy as np
 
-from pylops.optimization.base_linearoperator import BaseLinearOperator
+from pylops.linearoperator import BaseLinearOperator
 from pylops.signalprocessing import Convolve1D
 from pylops.utils._internal import _value_or_sized_to_array
 from pylops.utils.decorators import reshaped
